Call removeInterceptor() with instance
Previously removeInterceptor() only took an options object. This was
inconvenient if the client didn't wish to expose urls directly in their
tests.

removeInterceptor now also takes an instance of `Interceptor` and uses
the instance variables `baseUrl` and `_key` to filter for removal.

Added 3 test cases to cover http, https and regex path matching.