Do not abort a an already aborted request, do not emit an error twice
Fixes #882

When a request is aborted, "nock" emits an "abort"-error.
If the client reacts on that error by aborting the request (this is what
"popsicle" does), the code will end in an endless loop.
This commit makes that the the "abort"-code is only executed, if the
request has not already been aborted.