Have aborts follow the published API
I had some difficulty with this code:

    httpRequest.on('timeout', function () {
      trackFailure('timeout')
      httpRequest.abort()
    })

    httpRequest.on('error', function (err) {
      // Count error if request wasn't aborted due to timeout
      if (!httpRequest.aborted) {
        trackFailure(err)
      }
      callback(err)
    }

When `nock` was not in use, `trackFailure()` would be called once on
timeout.  With `nock` enabled, `trackFailure()` would be called twice
because `httpRequest.aborted` would be undefined.

This change creates and uses the `aborted` property in the same
fashion as the system `http` module.