docs(CODE_OF_CONDUCT.md): Change maintainer support email
I have changed the maintainer.io support email from `richard@maintainer.io` to `coc+nock@maintainer.io` in an attempt to make it more applicable in other CoCs, less tied to my own namespace, and to point out that emails which go to this email do not necessarily go to me, but to my team, which may at times include more than just a single perceived cis white male. Hopefully, this change will make it easier to email maintianer as a third party in disputes.