docs: Add a section on becoming a Maintainer
This section should represent ways to help out as a contributor, and what we look for in maintainers. I tried to highlight the main things that community members can do to help out, and what the maintainers look for when we are thinking about having someone join the project. What do you think?