test(test_intercept.js): Remove pointless reqheaders test
After studying this test for a good hour with @RichardLitt and @jlilja, we determined that it wasn’t asserting anything useful. There are other tests of `reqheaders` and we were unable to infer what this was intending to test. So, remove this one.

Add clarifying comments in other tests we studied along the way.