test(test_aws_dynamo): Remove AWS Dynamo functional test
This reverts 90662fc1fcc74586b157dddd5a408a692579ff20 which added a wide-bracket functional test for #223.

I appreciate the goal of defensively avoiding a regression. I could imagine circumstances where it’s worth installing some prominent client library to ensure nock keeps working with some complex workflow.

However the default when finding a bug in nock should be to isolate the bug in a small test case, and continue to maintain that small example to protect from regressions.

In this case there wasn’t even a bug.

There’s another regression test for #256 which also uses `aws-sdk` but that’s going to take a bit more effort to rework.