fix: Restore behavior of Interceptor.filteringPath (#1543)
Calling `filteringPath` on the intercept instance was broken as the transform fn set on the scope had the wrong name. Proxying to the Scope’s method allows for the regex version to work too. The bulk of the changed lines come from moving the tests to their appropriate file since the real logic acts on the Scope.

Found when looking at Uncovered lines in coveralls.