CR feedback. Separate two interceptor reply fns.
Use different attribute names for the reply functions that return a body
vs return an array for the whole response (dynamic).

Rejigger RequestOverrider.end to return early based which flow is used.
It would be nice if this method was split up into more pure functions
to reduce the complexity.