docs: Remove unnecessary uses of filteringRequestBody. (#1578)
Before v2.0, it was not possible to **not** filter on request bodies.

For [users wanting](https://github.com/nock/nock/issues/39) to match regardless of the request body, a [convention](https://github.com/nock/nock/commit/460aac919149fb190e910117c10c82f9d0e4c68e)
became the norm of using `filteringRequestBody` to modify the body into
a special string (`*`) then match against that specific string in the
Interceptor.

In more recent versions, the preferred way to not match request bodies
is to not provide a body argument when creating an Interceptor. This
updates the docs to reflect that.