test: fix flappy nock back test (#1595)
- Change fixture filename to be different than the one used in
`test_back_2.js` to avoid race conditions.

- Update the server response to have higher resolution entropy.

- Minor refactoring and typos.

It's unclear which systems this flap is prevalent on, but on my macOS
Mojave it would error at least 10% of the time. I narrowed it down to
the synchronous file system calls not being 100% safe of race conditions
when two test suites were creating, updating and deleting the same file
in a short period of time.