fix: allow unmocked when providing literal search params. (#1614)
fixes: #1421

There was inconsistent behavior around how search params were handled
when they were provided as part of a URI string to new Interceptors.
The issue happen to come to light when `allowUnmocked` was set.

The fix normalizes the behavior by refactoring the constructor of
Interceptor to look for literal search params. If present, they're
stripped from the `path` attribute and set via the `query` method.

As part of the work, `Interceptor.query` was modified to accept
`URLSearchParams` instances as valid input.