refactor: Correct the inverted logic of isUtf8Representable (#1752)
In 9da400f98aa2d7fe06e813b01403877fed4f2065 – which isn’t linked to the PR for some reason –`isBinaryBuffer()` was given a better name `isUtf8Representable`. The rename should have also inverted the logic, as a so-called “binary buffer” was one that was _not_ representable in UTF-8.

This completes the refactor, replacing all references to “binary” with “utf8-representable”. That is, except the parameter to `nock.define()`, which is still using `responseIsBinary` which has been deferred.