refactor: Mocha DSL for test_reply_* (#1773)
This adds a fair amount of nesting to these tests. Nesting is a mixed bag. It makes the tests a little harder to work with visually, so that's a downside. Though it also helps to clarify the specific behaviors being tested, and makes it much easier to reason about the collection of behaviors as a whole.