Remove needle tests
I studied #146 #175 and #193 for a while, but don't understand exactly what I'd need to do to replicate the problem from #146 without using needle. I also can't quite make sense of the code changes made to support needle, as it's all changed a lot since this test was added. (It goes back to 2014!)

I don't relish removing the test without having an isolated test to replace it with, so I don't mind leaving this around a little while if someone wants to try to do that. At the same time, tests like this are completely unmaintainable and really need to go. So another option, if someone feels passionate about preventing this particular regression is to pursue #1776.

To be clear, this bug seemed not specific to needle.