feature: only activate once, when on
While debugging the Jest memory leak, I was surprised to find the Nock
was "activating" twice when imported. This is because intercept.js
was calling `activate()` directly _AND_ back was setting the mode, which
calls `restore()` and `activate()`. There is no need for this double
initialization. Think of the CPUs we could be saving.
I also discovered that Nock "activates" even if the proc is started with
the `NOCK_OFF` env var set, which means Nock is polluting the global
scope even when callers have asked it not to.

Since it felt like the `back.setMode()` call was hiding in its module,
I've moved it to the entry index file and wrapped it in a `isOn` check.