Temporarily disable Node 8 CI to unblock release
CI is broken for Node 8 beacuse semantic-release won't install. This isn't just breaking all the builds, it's also blocking release.

We're dealing with updating the CI this in #1870 and maybe we'll be able to fix this. If not, we should also prioritize #1871 so we don't accidentally ship a breaking change for Node 8.

I'm removing this for now, so we can get the patch for #1836 shipped.