fix(router): stall finish event to avoid dup
Delays the emitting of the "finish" event on the request until it knows for sure an Interceptor is going to playback.
This stops a double emit for unmocked requests.

Also removes the "end" event from firing on the request. This event was added in the first commit with code for Nock
and was used to kicked off the internals that now live in `startPlayback`.
https://github.com/nock/nock/commit/31a500c96a9a5329503c2968b5d5f87c0be5a4f8#diff-503754c07b5098227646f913eee85885R27-R76

closes #1832