fix: Parallel requests on same Interceptor are exposed correctly in reply functions.
Fixes regression created in https://github.com/nock/nock/pull/2033.
Reply functions get access to the current request via the Interceptor,
which acts as the context during the function executions.
The previous change, incorrectly moved the line where the request was attached
to the Interceptor up in the flow and out of the same microtask that calls the
reply functions. Causing requests in parallel to lose the reference to the correct request.