test(recorder): skip requests from previous sessions
recorder.js has a block of code that skips storing the output of
requests initiated during another recording session.
The coverage for this block was not being covered when the tests where
run via Mocha and only covered via Tap by coincidence.
'http request without callback should not crash' was the culprit of the
coincidence via Tap as it never properly waited for the request to end
before cleaning up. Therefore subsequent tests would start a new
recording session and enter the block of code in question.
That test was updated to wait until the response has ended before
calling `done` on the test.
Then a new test was added to explicitly test the logic of skipping
recording of requests from previous sessions.