docs: add note on removeInterceptor usage (#2035)
The fact that `removeInterceptor` works on some chains and doesn't work on other chains is very confusing because it means that existing code may break while the behavior is getting only extended (opposite to modified). Also, because documentation is not giving a clear difference on what chain methods returns what type of values. There are a bunch of issues on it:
https://github.com/nock/nock/issues/1822
https://github.com/nock/nock/issues/1117
https://github.com/nock/nock/issues/600