feat(intercept): path matching only uses pathname
BREAKING CHANGE: The query/search portion of the path is now always excluding during path matching.

Previously, the pathname and search string  were included during path comparison by default, however, the search string was removed during the matching step if the `query` method had been called or a search string was provided as part of a string to the first argument of a "verb" method.
This change makes the experience consistent by making query/search matching exclusively opt-in.